+++ ScholarNet/app/Http/Controllers/AssignmentSubmissionsController.php
@@ -2,18 +2,22 @@
 
 namespace App\Http\Controllers;
 
+use Carbon\Carbon;
 use App\Models\Assignment;
-use App\Models\assignment_submissions;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
-use Carbon\Carbon;
+use Illuminate\Support\Facades\Gate;
+use App\Models\assignment_submissions;
+
 class AssignmentSubmissionsController extends Controller
 {
     public function create(Assignment $assignment){
+        Gate::authorize('assignment-details',$assignment);
         return view('student/submitmyassignment',compact('assignment'));
     }
 
     public function store(Assignment $assignment,Request $request){
+        Gate::authorize('assignment-details',$assignment);
         $studentId = auth()->user()->id;
 
 
@@ -49,6 +53,7 @@ class AssignmentSubmissionsController extends Controller
     }
 
     public function destroy(assignment_submissions $exercice1){
+        Gate::authorize('exercice',$exercice1);
         $exercice = assignment_submissions::find($exercice1->id);
         $exercice->delete();
         //dd($exercice);
@@ -56,10 +61,13 @@ class AssignmentSubmissionsController extends Controller
     }
 
     public function download(assignment_submissions $exercice){
+        Gate::authorize('exercice',$exercice);
         return response()->download('storage/'.$exercice->fichier);
     }
     public function submissions($assignmentId)
     {
+        $assignment = Assignment::findOrFail($assignmentId);
+        Gate::authorize('assignment-details',$assignment);
         $submissionA = DB::table('assignment_student')
         ->where('assignment_id', $assignmentId)
         ->get();
@@ -70,6 +78,8 @@ class AssignmentSubmissionsController extends Controller
 
     public function updateNote($assignment_id, $student_id, Request $request)
     {
+        $assignment = Assignment::findOrFail($assignment_id);
+        Gate::authorize('assignment-details',$assignment);
         DB::table('assignment_student')
             ->where('assignment_id', $assignment_id)
             ->where('student_id', $student_id)
